notion-be/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==39.0.2
pytest==7.3.1
httpx==0.24.0
pytest-asyncio==0.21.0
email-validator==2.0.0 